Title: 숫자들의 합이 7의 배수, Time: 11ms, Memory: 1MB, Status: Passed - CodeTree

+++ 250213/숫자들의 합이 7의 배수/the-sum-of-the-numbers-is-a-multiple-of-7.cpp
@@ -3,26 +3,26 @@
 
 using namespace std;
 
-int n;
-int arr[50001];  // 최대 50000개
-unordered_map<int, int> mod_idx; // 나머지 값 -> 최초 등장 인덱스 저장
+long long n;
+long long arr[50001];  // 최대 50000개
+unordered_map<long long, long long> mod_idx; // 나머지 값 -> 최초 등장 인덱스 저장
 
 int main() {
     ios::sync_with_stdio(false);
     cin.tie(0);
 
     cin >> n;
-    for (int i = 1; i <= n; i++) {
+    for (long long i = 1; i <= n; i++) {
         cin >> arr[i];
     }
 
-    int prefix_sum = 0;
-    int max_len = 0;
+    long long prefix_sum = 0;
+    long long max_len = 0;
     mod_idx[0] = 0; // 누적합이 0일 경우 대비
 
-    for (int i = 1; i <= n; i++) {
+    for (long long i = 1; i <= n; i++) {
         prefix_sum += arr[i];
-        int mod_val = prefix_sum % 7;
+        long long mod_val = prefix_sum % 7;
 
         // 나머지 값이 처음 등장하면 저장
         if (mod_idx.find(mod_val) == mod_idx.end()) {
